requirements.txt
streamlit
pandas
numpy
scikit-learn
nltk
pdfplumber
python-docx
sentence-transformers
torch